Converted scrollTo tests to intern.
This is not a one to one conversion. The original scrollTo tests were
hard to convert to intern. These tests run through all of the document
layout scenarios and verify that the target element is scrolled into the
viewport.

The target element is considered in view if the top or bottom and the
right side of the target are within the viewport.